Remove circular dependency introduced in the last release.
