+++ src/components/popup/tokens/PriceChart.tsx
@@ -11,7 +11,7 @@ import { Flex } from "~components/common/Flex";
 import { PriceChartModal } from "./PriceChartModal";
 import { LineChart } from "../chart/LineChart";
 import { useTokenPrice } from "~tokens/hooks";
-import { AR_PROCESS_ID, AO_PROCESS_ID } from "~tokens/aoTokens/ao";
+import { AR_PROCESS_ID, AO_PROCESS_ID } from "~tokens/aoTokens/ao.constants";
 
 interface PriceChartProps {
   symbol: CoinGeckoSymbol;
