+++ src/routes/popup/send/index.tsx
@@ -16,9 +16,10 @@ import { searchArNSName } from "~lib/arns";
 import SliderMenu from "~components/SliderMenu";
 import { useLocation } from "~wallets/router/router.utils";
 import { ExtensionStorage, TempTransactionStorage } from "~utils/storage";
-import { AR_PROCESS_ID, type TokenInfo } from "~tokens/aoTokens/ao";
+import type { TokenInfo } from "~tokens/aoTokens/ao";
 import { useStorage } from "@plasmohq/storage/hook";
 import { NoAvatarIcon } from "~components/Avatar";
+import { AR_PROCESS_ID } from "~tokens/aoTokens/ao.constants";
 
 // default size for the qty text
 export const arPlaceholder: TokenInterface = {
